:art: refactor: refactor word filter to their own file instead of word interface.ts
