+++ src/app/shared/models/word.interface.ts
@@ -25,11 +25,6 @@ export interface WordSense {
   globalIndex?: number;
 }
 
-export interface WordWithSense {
-  word: Word;
-  wordSenses: WordSense[];
-}
-
 export interface WordWithAttemptsAndSuccess {
   word: Word;
   attempts: number;
@@ -42,70 +37,4 @@ export interface WordSenseInfoWithoutWord {
   attempts: number;
   success: number;
   categories: Category[];
-}
-
-export interface WordFilterOptions {
-  partSpeeches: PartSpeech[];
-  levels: Level[];
-  categories: Category[];
-  persons: Person[];
-  genders: Gender[];
-  numbers: Number[];
-  moodWithTenses: MoodWithTense[];
-}
-
-export interface WordFilterRequest {
-  textFiltered?: string[];
-  minAccuracy?: number;
-  maxAccuracy?: number;
-  partSpeeches?: PartSpeech[];
-  levels?: Level[];
-  categories?: Category[];
-  persons?: Person[];
-  genders?: Gender[];
-  numbers?: Number[];
-  moodWithTenses?: MoodWithTense[];
-  tenses?: Tense[];
-}
-
-export function compareWordFilterRequest(
-  wwordFilterRequest1: WordFilterRequest,
-  wwordFilterRequest2: WordFilterRequest,
-) {
-  return (
-    wwordFilterRequest1.minAccuracy === wwordFilterRequest2.minAccuracy &&
-    wwordFilterRequest1.maxAccuracy === wwordFilterRequest2.maxAccuracy &&
-    compareArrays(
-      wwordFilterRequest1.textFiltered,
-      wwordFilterRequest2.textFiltered,
-    ) &&
-    compareArrays(
-      mapArrayByProperty(wwordFilterRequest1.partSpeeches, 'code'),
-      mapArrayByProperty(wwordFilterRequest2.partSpeeches, 'code'),
-    ) &&
-    compareArrays(
-      mapArrayByProperty(wwordFilterRequest1.levels, 'code'),
-      mapArrayByProperty(wwordFilterRequest2.levels, 'code'),
-    ) &&
-    compareArrays(
-      mapArrayByProperty(wwordFilterRequest1.categories, 'code'),
-      mapArrayByProperty(wwordFilterRequest2.categories, 'code'),
-    ) &&
-    compareArrays(
-      mapArrayByProperty(wwordFilterRequest1.persons, 'code'),
-      mapArrayByProperty(wwordFilterRequest2.persons, 'code'),
-    ) &&
-    compareArrays(
-      mapArrayByProperty(wwordFilterRequest1.genders, 'code'),
-      mapArrayByProperty(wwordFilterRequest2.genders, 'code'),
-    ) &&
-    compareArrays(
-      mapArrayByProperty(wwordFilterRequest1.numbers, 'code'),
-      mapArrayByProperty(wwordFilterRequest2.numbers, 'code'),
-    ) &&
-    compareArrays(
-      mapArrayByProperty(wwordFilterRequest1.tenses, 'code'),
-      mapArrayByProperty(wwordFilterRequest2.tenses, 'code'),
-    )
-  );
 }
